Add DataSet to wrap some stuff in
