--- Sources/PerformanceTesting/LinearRegression.swift
@@ -1,54 +1,0 @@
-//
-//  LinearRegression.swift
-//  PerformanceTesting
-//
-//  Created by James Bean on 8/10/17.
-//
-
-import Darwin
-
-internal struct Regression {
-    public let slope: Double
-    public let intercept: Double
-    public let correlation: Double
-}
-
-/// Calculates the linear regression of the given `data`.
-internal func linearRegression(_ data: [(Double,Double)]) -> Regression {
-    let xs = data.map { $0.0 }
-    let ys = data.map { $0.1 }
-    let sumOfXs = xs.sum
-    let sumOfYs = ys.sum
-    let sumOfXsSquared = xs.map { pow($0,2) }.sum
-    let sumOfXsTimesYs = data.map(*).sum
-
-    let denominator = Double(data.count) * sumOfXsSquared - pow(sumOfXs, 2)
-    let interceptNumerator = sumOfYs * sumOfXsSquared - sumOfXs * sumOfXsTimesYs
-    let slopeNumerator = Double(data.count) * sumOfXsTimesYs - sumOfXs * sumOfYs
-    let intercept = interceptNumerator / denominator
-    let slope = slopeNumerator / denominator
-
-    return Regression(
-        slope: slope,
-        intercept: intercept,
-        correlation: correlation(data, sumOfXs: sumOfXs, sumOfYs: sumOfYs, slope: slope)
-    )
-}
-
-/// Helper function to calculate the regression coefficient ("r") of the given dataset.
-private func correlation(
-    _ data: [(Double,Double)],
-    sumOfXs: Double,
-    sumOfYs: Double,
-    slope: Double
-) -> Double
-{
-    let meanOfYs = sumOfYs / Double(data.count)
-    let squaredErrorOfYs = data.map { pow($0.1 - meanOfYs, 2) }.sum
-    let denominator = squaredErrorOfYs
-    guard denominator != 0 else { return 0 }
-    let meanOfXs = sumOfXs / Double(data.count)
-    let squaredErrorOfXs = data.map { pow($0.0 - meanOfXs, 2) }.sum
-    let numerator = squaredErrorOfXs
-    return sqrt(numerator / denominator) * slope
-}
